rename: move inner class to upper level & move package: test -> main (#15)
